added gripper mixing
